values.js moved to about folder + changed name to values_array.js
